Move passToPackman from config.go to cmd.go
