+++ config.go
@@ -195,22 +195,3 @@ func continueTask(s string, def string) (cont bool) {
 
 	return true
 }
-
-// PassToPacman outsorces execution to pacman binary without modifications.
-func passToPacman(parser *argParser) error {
-	var cmd *exec.Cmd
-	args := make([]string, 0)
-	
-	if parser.needRoot() {
-		args = append(args, "sudo")
-	}
-	
-	args = append(args, "pacman")
-	args = append(args, parser.formatArgs()...)
-	
-	cmd = exec.Command(args[0], args[1:]...)
-
-	cmd.Stdin, cmd.Stdout, cmd.Stderr = os.Stdin, os.Stdout, os.Stderr
-	err := cmd.Run()
-	return err
-}
